Added images of frames to stacked plotter
I added a very rudimentary feature that adds images of the frames with and without ROIs.  This this requires substantial modification to make it more ease-of-access (i.e. preventing it from crashing if someone doesn't specify where the pictures are).  Also Renan made some changes to the denoising functions while he was working on my computer, but I'm not sure what those were.