Started writing nerve recording function
Created a new script "extractTBN" to read the extracellular nerve data associated with each recording.  The data are stored in .tbn files, which are a type of binary data files (BNC).  So far, I have written a script that successfully reads the files and converts the data into double formats.  Additionally, I have found that the Turbo-SM64 manual is accurate in describing how to acquire "header" information, which reads the BNC data as int16.  From there, I can accurately index the number of channels I recorded from.  Additionally, as the Turbo-SM64 manual specified, the data should be extracted as float64, and diving the length of the resulting vector of doubles by the number of channels results in 1:1 nerve:VSD data.  I.e. there are 61110 VSD frames, and 61110 nerve points per channel.  I explain each of these steps because I believe I successfully read the data and identified what data is what.  However, I am describing all of this detail because the sizes of the values in the resulting matrix are just so enormous, that they nearly exceed the limits of MATLAB.  I need to somehow size-down the data, because it looks really crazy.  TL; DR: it does not work yet