Update Intan_gui.m
fixed a few bugs.  The x-axis was disappearing when removing bottom axis.  A bug still persists, the ROI image becomes stretched when opening a file after a previously opening a file, even if you open the same file.  It is changing the xlimits of the axis to the same limits of the traces in line 696, it is also removing the text prior to line 696.